requirements.txt
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
docutils==0.17.1
idna==2.10
importlib-metadata==4.5.0
jeepney==0.6.0
keyring==23.0.1
kiwisolver==1.3.1
lib==3.0.0
lxml==4.6.3
matplot==0.1.9
matplotlib==3.4.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandas-datareader==0.9.0
Pillow==8.2.0
pkginfo==1.7.0
pycparser==2.20
Pygments==2.9.0
pyloco==0.0.139
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
SimpleWebSocketServer==0.1.1
six==1.16.0
tqdm==4.61.0
twine==3.4.1
typing==3.7.4.3
urllib3==1.26.5
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.0.1
zipp==3.4.1
